Prior exposures calculated by sessions table only
